requirements.txt
Django==3.0.2
gunicorn==19.9.0
line-bot-sdk==1.12.1
dj-static
dj-database-url
psycopg2
static3
requests